+++ app/src/main/kotlin/com/pitchedapps/frost/injectors/CssHider.kt
@@ -9,7 +9,7 @@ import android.webkit.WebView
  */
 enum class CssHider(vararg val items: String) : InjectorContract {
     CORE("[data-sigil=m_login_upsell]", "role=progressbar"),
-//    HEADER("#header", "[data-sigil=MTopBlueBarHeader]",
+    //    HEADER("#header", "[data-sigil=MTopBlueBarHeader]",
 //            "#header-notices", "[data-sigil*=m-promo-jewel-header]"),
     ADS("article[data-xt*=sponsor]",
             "article[data-store*=sponsor]"),