+++ app/src/main/kotlin/com/pitchedapps/frost/settings/Debug.kt
@@ -10,6 +10,7 @@ import com.pitchedapps.frost.activities.DebugActivity
 import com.pitchedapps.frost.activities.SettingsActivity
 import com.pitchedapps.frost.activities.SettingsActivity.Companion.ACTIVITY_REQUEST_DEBUG
 import com.pitchedapps.frost.debugger.OfflineWebsite
+import com.pitchedapps.frost.facebook.FB_URL_BASE
 import com.pitchedapps.frost.facebook.FbCookie
 import com.pitchedapps.frost.facebook.FbItem
 import com.pitchedapps.frost.parsers.FrostParser
@@ -90,18 +91,12 @@ private fun Context.createEmail(parser: FrostParser<*>, content: Any?) =
 
 private const val ZIP_NAME = "debug"
 
-fun SettingsActivity.sendDebug(urlOrig: String) {
-
-    val url = when {
-        urlOrig.endsWith("soft=requests") -> FbItem.FRIENDS.url
-        urlOrig.endsWith("soft=messages") -> FbItem.MESSAGES.url
-        urlOrig.endsWith("soft=notifications") -> FbItem.NOTIFICATIONS.url
-        urlOrig.endsWith("soft=search") -> "${FbItem._SEARCH.url}?q=a"
-        else -> urlOrig
-    }
+fun SettingsActivity.sendDebug(url: String, html: String?) {
 
     val downloader = OfflineWebsite(url, FbCookie.webCookie ?: "",
-            DebugActivity.baseDir(this))
+            baseUrl = FB_URL_BASE,
+            html = html,
+            baseDir = DebugActivity.baseDir(this))
 
     val md = materialDialog {
         title(R.string.parsing_data)