+++ docs/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.8.3
+* Add full notification channel support
+* Fix sound spam for multiple notifications
+
 ## v1.8.2
 * Fix duplicate notification sounds
 * Fix map redirecting to blackberry